requirements.txt
aiohttp==3.7.4.post0
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.5
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
flask-restplus==0.13.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
pandas==1.1.4
passlib==1.7.4
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.4
pylint==2.6.0
pymongo==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pythainlp==2.2.6
python-barcode==0.13.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2020.4
qrcode==6.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
tinydb==4.3.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.2
virtualenv==20.1.0
Werkzeug==0.16.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.8
xmltodict==0.12.0
yarl==1.6.3
yattag==1.14.0
zipp==3.4.0
